Rename blog post for clarity
The blog post originally named 2023-08-26.md has been renamed to 2023-08-26-Lowest-Common-Ancestor-of-a-Binary-Search-Tree.md. This was done to provide clarity on the specific content of the blog post, which focuses on solving the Lowest Common Ancestor problem in a Binary Search Tree. The new filename includes the specific topic, allowing for easier navigation and identification of the post's subject matter.