Remove duplicated titles in blog posts
In this commit, redundant titles were removed from three blog posts: "discuss-test-cases-before-test-generation", "Lowest-Common-Ancestor-of-a-Binary-Search-Tree", and "BlackBox-Unit-Test". Initially, the titles were present both in the post names and in the h1 headings inside the content. For better readability and coherent structure, the titles inside the body were removed. The title defined by the post filename is enough to identify the subject matter of the article.