Update diagram terminology in blog post
Updated the diagram for the 'Blackbox Testing Validate Binary Search Tree' blog post. The term 'Human' was changed to 'Actor' in order to align terminology with industry standards and ensure consistency throughout the blog. This is important in ensuring that readers, especially those new to the industry, understand and can follow the concepts described in the blog post.