Add .idea and chats folders to Jekyll ignore list
Added .idea and chats directories to the Jekyll ignore list in _config.yml file. This change will prevent Jekyll from processing these directories when building the site. The .idea directory is specific to JetBrains IDEs and should not be included in the build process. The chats directory contains private data irrelevant to the site content.