Update contact information with LinkedIn and Stack Overflow URLs
In this commit, we have uncommented and filled in the LinkedIn and Stack Overflow sections in the contact.yml config file. The specific URLs for user 'jegors-cemisovs' on both platforms have been added. This will provide visitors with more ways to interact and get in touch with the site owner.