Change blog layout type and add blockquote on test.md
Changed 'layout: post' to 'layout: home' in test.md to give the file a home page layout. This change serves to improve content presentation. Also added a blockquote at the end of the file to accommodate well-known quotes, to give more dimension to the content, enhance readability and enrich the user's experience.