Refactor dialogue formatting and improve Java code wrapping in blog post
This commit combines the dialogue into a single line for readability purposes. In addition, `{% raw %}` and `{% endraw %}` tags have been added to the Java code blocks in the blog post on black box testing for "Set Matrix Zeroes". This prevents Liquid, the templating engine in Jekyll, from processing the code inside. The blog post filename also gets a minor tweak for more clarity.