Add clarity to post titles, tags, and categories
This commit adds clarity and uniformity to posts by ensuring the consistent use of 'BlackBox Testing:' and 'WhiteBox Testing:' in titles. 'black-box' and 'white-box' in tags have been replaced by the non-hyphenated variants. Categories and tags for certain posts have been updated to better reflect their content. A new post focussed on rotating an image has also been added. A tweak to the pin status of the post '2023-08-26-whitebox-testing-lowest-common-ancestor-of-a-binary-search-tree.md' ensures it appears at the top of relevant lists. Additionally, filenames of posts related to binary tree inorder traversal have been edited for improved comprehension.