Update post metadata and improve categorization
This commit refines post categorization and tagging, better aligning them with their content. It ensures consistency in the use of 'BlackBox Testing:' and 'WhiteBox Testing:' in post titles and in the tags. The commit also deletes the 'mermaid: true' metadata which was unused, thus streamlining the metadata of posts. This commit results in adding a new post focusing on image rotation in the context of black-box testing.