Update post title in whitebox-testing-spiral-matrix.md
Changed the title of the post from "WhiteBox Testing: Spiral Matrix" to "The problem of choosing the proper annotation. MethodSource vs. CsvSource". The update was necessary to better reflect the content of the post and to ensure it is more informative and directly related to the discussed topics, primarily focusing on comparing two different types of annotations in Java: MethodSource and CsvSource.