Add new informative post about Java Developer: FlipFlopPredicate
Added a new post which gives a detailed explanation about the implementation of a Generic FlipFlopPredicate Class in Java. This includes information about the private variables, constructor and test method of the class. This post will assist Java developers to understand and implement this efficiently in their own code, adhering to the latest best practices and clean code principles.