Refactor post titles and filenames to be more descriptive
The specific post filenames and their respective titles for 'FlipFlopPredicate' have been refactored to be more descriptive and informative. Name changes indicate that each file is a dedicated post discussing the topic more from a certain role perspective (QA Engineer, Business Analyst and Java Developer). Hence, the filenames and titles are changed from the generic 'BDD' and 'BA' to the full 'qa-engineer' and 'business-analyst', providing a better understanding of the context for the audience.