Update Giscus configuration in _config.yml
The Giscus configuration was adjusted in the _config.yml file. This was done to change the commenting settings on the blog. Specifically, the repo has been updated to "rabestro/blog-discussions" and the mapping set to 'title'. Moreover, the position of the comment input box has been moved to the top. This provides better visibility and accessibility for users wishing to leave comments.