Add new blog post about FizzBuzz and filtering in Java
Created a new blog post named '_posts/2023-10-26-fizzbuzz-and-the-art-of-filtering.md'. The blog post covers the topic of the classic FizzBuzz problem and its relation to filtering with the Stream API in Java. It includes several tasks that employ predicates to achieve solutions. Use of predicates in Java and their flexibility is showcased extensively throughout the post and many solutions are presented, including those using the 'and', 'or', 'negate' methods from the 'IntPredicate' interface as well as the usage of logical operators. The post aims to inspire developers to explore the vast possibilities that predicates offer in functional programming in Java.