Remove outdated blog post 'Beyond FizzBuzz'
The blog post 'Beyond FizzBuzz' has been deleted. This article was no longer relevant or accurate due to developments in Java's Stream API and updates to the FizzBuzz algorithm problem.

The blog post 'FizzBuzz and the art of filtering' is a more comprehensive article that not only covers the information mentioned in 'Beyond FizzBuzz' but also showcases advanced utilization of filtering in Stream API in a much more precise way. Thus, to avoid redundancy and confusion among readers, the outdated post is removed. Additionally, some irrelevant bold format in the 'FizzBuzz and the art of filtering' has been cleared for a cleaner look.