Add method to check if user is payable and its test cases
Added a new Service interface that includes a static method 'isPayableUser' and a predicate class 'IsPayableUserPredicate'. The method checks if a user is over 18 years old using a BDD approach. Birthday and user nullability are handled. A constant 'LEGAL_AGE' is defined, improving the readability. Fixed a flaw where the method was failing for users who are exactly 18 years old. Furthermore, included test cases for scenarios such as user is exactly 18 years old, younger, older and null. Additionally, a test case for null birthday was added.